feat(client): added usePhone hook
