+++ strife-client/src/types/socket-types.js
@@ -1,3 +1,1 @@
-export const CREATE = 'CREATE';
-export const EMIT = 'EMIT';
-export const LISTEN = 'LISTEN';
+export const UPDATE_SOCKET = 'UPDATE_SOCKET';
